trunner: utils: rename from wrappers, change class names
JIRA: PD-268
