+++ trunner/runners/ARMV7M7Runner.py
@@ -15,7 +15,7 @@ from pexpect.exceptions import TIMEOUT, EOF
 from trunner.tools.color import Color
 from .common import LOG_PATH, PloError, PloTalker, DeviceRunner, RebootError
 from .common import GPIO, rootfs
-from .wrappers import Phoenixd, PhoenixdError, append_output
+from .utils import Phoenixd, PhoenixdError, append_output
 from .flasher import NXPSerialFlasher
